Conflicts:
	FastSimulation/Tracking/python/IterativeInitialStep_cff.py

+++ DQM/EcalCommon/src/MESetBinningUtils2.cc
@@ -317,7 +317,7 @@ namespace ecaldqm
           break;
         case kProjEta:
           specs.nbins = nEBEtaBins + 2 * nEEEtaBins;
-          specs.edges = new double[specs.nbins + 1];
+          specs.edges = new float[specs.nbins + 1];
           for(int i(0); i <= nEEEtaBins; i++)
             specs.edges[i] = -3. + (3. - etaBound) / nEEEtaBins * i;
           for(int i(1); i <= nEBEtaBins; i++)
